delete alien/party verification

+++ src/Service/alien.js
@@ -1,4 +1,6 @@
 import AlienRepository from "../Repository/alien.js";
+import RegisterRepository from "../Repository/register.js";
+import BacklogRepository from "../Repository/backlog.js";
 
 async function createAlien(alien) {
   return await AlienRepository.insertAlien(alien);
@@ -16,8 +18,19 @@ async function updateAlien(alien) {
   return await AlienRepository.updateAlien(alien);
 }
 
-//TODO only if there is no costumer with this party id
 async function deleteAlien(id) {
+  const register = await RegisterRepository.getRegisterByAlienId(id);
+  if (register.length > 0) {
+    throw new Error("Esse alien está dentro de uma festa");
+  }
+
+  const backlog = await BacklogRepository.getBacklogByAlienId(id);
+  if (backlog.length > 0) {
+    throw new Error(
+      "Esse alien tem backlog registrado, e por isso não é possivel excluir"
+    );
+  }
+
   return await AlienRepository.deleteAlien(id);
 }
