+++ src/Service/party.js
@@ -1,4 +1,6 @@
 import PartyRepository from "../Repository/party.js";
+import RegisterRepository from "../Repository/register.js";
+import BacklogRepository from "../Repository/backlog.js";
 
 async function createParty(party) {
   return await PartyRepository.insertParty(party);
@@ -17,7 +19,18 @@ async function updateParty(party) {
 }
 
 async function deleteParty(id) {
-  //TODO only if there is no costumer with this party id
+  const register = await RegisterRepository.getRegisterByPartyId(id);
+  if (register.length > 0) {
+    throw new Error("existe alien dentro dessa festa");
+  }
+
+  const backlog = await BacklogRepository.getBacklogByPartyId(id);
+  if (backlog.length > 0) {
+    throw new Error(
+      "Essa balada tem backlog registrado, e por isso não é possivel excluir"
+    );
+  }
+
   return await PartyRepository.deleteParty(id);
 }
